requirements.txt
django>3
djangorestframework
